+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   - 配置仓库 -> 配置节点编辑器新增文件列表节点子编辑面板。
 
 - `component-api` 子模功能新增。
+  - 新增 `src/api/settingrepo/navigationNode.ts` API 接口。
+  - 更新 `src/api/settingrepo/settingNode.ts` API 接口。
   - 新增 `src/api/settingrepo/fileListNode.ts` API 接口。
   - 新增 `src/api/settingrepo/fileNode.ts` API 接口。